Update Genetic.java
Hi Mark, 
Thank you for this great code. I have been using a modified version of your 1996 code to demo GA's to my undergrad students.
If think the roulette wheel code should be based on numChromosomes. When you set a bit string length of 10 and a population of 20, the highest index you get in the roulette wheel is 9. This is because the num_trials is currently based on numGenesPerChromosome.
Regards
Stephen